+++ libs/angular-ui-kit/src/index.ts
@@ -34,6 +34,8 @@ export * from './lib/models/navigation-item.interface';
 // Accordion Component
 export * from './lib/accordion/accordion.component';
 export * from './lib/accordion/accordion-group.component';
+export * from './lib/accordion/accordion-header.directive';
+export * from './lib/accordion/accordion-content.directive';
 
 // Button Component
 export * from './lib/button/button.component';